# openapi.yaml
openapi: 3.0.0
info:
  description: This is a API for tourism to search destinations
  version: 1.0.0
  title: Tourism API
  contact:
    email: muhammadarash@gmail.com
servers:
  - url: https://apitourism.herokuapp.com
...
paths:
...
  /api/image/{destinationUUID}:
    post:
      tags:
        - image
      summary: Add an image for a destination
      description: This can only be done by the logged in admin
      parameters:
        - name: destinationUUID
          in: path
          description: Destination uuid to add image of destination
          required: true
          schema:
            type: string
            format: byte
      responses:
        '200':
          description: Success adding an image
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    $ref: '#/components/schemas/Image'
        '400':
          description: Failed for Adding an image
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    $ref: '#/components/schemas/Error'
        '401':
          description: The requester was not an admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    $ref: '#/components/schemas/Error'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
      requestBody:
        description: File image of destination
        required: true
        content:
          image:
            schema:
              type: string
              format: binary
  /api/image/{imageUUID}:
    delete:
      tags:
        - image
      summary: Delete image of destinations by uuid
      description: This can only be done by the logged in admin
      parameters:
        - name: imageUUID
          in: path
          description: ID of the image that needs to be deleted
          required: true
          schema:
            type: string
            format: byte
      responses:
        '200':
          description: Success deleting an image
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    type: object
                    nullable: true
        '400':
          description: Failed for deleting an image
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    $ref: '#/components/schemas/Error'
        '401':
          description: The requester was not an admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
  /api/user/register:
    post:
      tags:
        - user
      summary: Create user
      responses:
        '200':
          description: Success registering an account
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    $ref: '#/components/schemas/UserRegistered'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    $ref: '#/components/schemas/Error'
      requestBody:
        $ref: '#/components/requestBodies/UserRegistered'
  /api/user/login:
    post:
      tags:
        - user
      summary: Logs user into the system
      responses:
        '200':
          description: Success logging in account
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    $ref: '#/components/schemas/UserLogged'
        '400':
          description: Failed for logging in account
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    $ref: '#/components/schemas/Error'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    $ref: '#/components/schemas/Error'
      requestBody:
        $ref: '#/components/requestBodies/UserLogged'
  /api/user/{userUUID}:
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in admin
      parameters:
        - name: userUUID
          in: path
          description: The user uuid that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success deleting a user
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    type: object
                    nullable: true
        '400':
          description: Failed for deleting a user
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    $ref: '#/components/schemas/Error'
        '401':
          description: The requester was not an admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/Meta'
                  data:
                    $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
...
components:
  schemas:
    Meta:
      type: object
      properties:
        message:
          type: string
          format: byte
        code:
          type: integer
          format: int64
        status:
          type: string
          format: byte
...
    Image:
      type: object
      properties:
        id:
          type: string
          format: byte
        destination_id:
          type: string
          format: byte
        file_name:
          type: string
          format: byte
...
    UserRegistered:
      type: object
      properties:
        id:
          type: string
          format: byte
        name:
          type: string
          format: byte
        email:
          type: string
          format: byte
        role:
          type: string
          format: byte
    UserLogged:
      type: object
      properties:
        id:
          type: string
          format: byte
        name:
          type: string
          format: byte
        email:
          type: string
          format: byte
        role:
          type: string
          format: byte
        token:
          type: string
          format: byte
...
    UserRegisteredRequest:
      type: object
      properties:
        name:
          type: string
          format: byte
        email:
          type: string
          format: byte
        password:
          type: string
          format: byte
    UserLoggedRequest:
      type: object
      properties:
        email:
          type: string
          format: byte
        password:
          type: string
          format: byte
    Error:
      type: object
      properties:
        errors:
          type: string
          format: byte
  requestBodies:
...
    UserRegistered:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserRegisteredRequest'
      description: User object that needs to be added to the store
      required: true
    UserLogged:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserLoggedRequest'
      description: User object that needs to be logged in
      required: true
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT